--- a/README.md
+++ b/README.md
@@ -4,7 +4,7 @@ Ecosystem Plugin for [Blast](https://blast.io/) support in Ape.
 
 ## Dependencies
 
-- [python3](https://www.python.org/downloads) version 3.8 up to 3.11.
+- [python3](https://www.python.org/downloads) version 3.8 up to 3.12.
 
 ## Installation