--- a/ape_base/__init__.py
+++ b/ape_base/__init__.py
@@ -3,32 +3,32 @@ from ape.api.networks import LOCAL_NETWORK_NAME, ForkedNetworkAPI, NetworkAPI, c
 from ape_geth import GethProvider
 from ape_test import LocalProvider
 
-from .ecosystem import NETWORKS, Base, BaseConfig
+from .ecosystem import NETWORKS, Blast, BlastConfig
 
 
 @plugins.register(plugins.Config)
 def config_class():
-    return BaseConfig
+    return BlastConfig
 
 
 @plugins.register(plugins.EcosystemPlugin)
 def ecosystems():
-    yield Base
+    yield Blast
 
 
 @plugins.register(plugins.NetworkPlugin)
 def networks():
     for network_name, network_params in NETWORKS.items():
-        yield "base", network_name, create_network_type(*network_params)
-        yield "base", f"{network_name}-fork", ForkedNetworkAPI
+        yield "blast", network_name, create_network_type(*network_params)
+        yield "blast", f"{network_name}-fork", ForkedNetworkAPI
 
     # NOTE: This works for local providers, as they get chain_id from themselves
-    yield "base", LOCAL_NETWORK_NAME, NetworkAPI
+    yield "blast", LOCAL_NETWORK_NAME, NetworkAPI
 
 
 @plugins.register(plugins.ProviderPlugin)
 def providers():
     for network_name in NETWORKS:
-        yield "base", network_name, GethProvider
+        yield "blast", network_name, GethProvider
 
-    yield "base", LOCAL_NETWORK_NAME, LocalProvider
+    yield "blast", LOCAL_NETWORK_NAME, LocalProvider