--- a/pyproject.toml
+++ b/pyproject.toml
@@ -6,7 +6,7 @@ exclude = "build/"
 plugins = ["pydantic.mypy"]
 
 [tool.setuptools_scm]
-write_to = "ape_base/version.py"
+write_to = "ape_blast/version.py"
 
 # NOTE: you have to use single-quoted strings in TOML for regular expressions.
 # It's the equivalent of r-strings in Python.  Multiline strings are treated as
@@ -25,7 +25,7 @@ addopts = """
     --cov-report term
     --cov-report html
     --cov-report xml
-    --cov=ape_base
+    --cov=ape_blast
 """
 python_files = "test_*.py"
 testpaths = "tests"