--- a/pyproject.toml
+++ b/pyproject.toml
@@ -15,7 +15,7 @@ write_to = "ape_blast/version.py"
 
 [tool.black]
 line-length = 100
-target-version = ['py38', 'py39', 'py310', 'py311', 'py312']
+target-version = ['py39', 'py310', 'py311', 'py312']
 include = '\.pyi?$'
 
 [tool.pytest.ini_options]