--- a/setup.py
+++ b/setup.py
@@ -60,11 +60,11 @@ setup(
     url="https://github.com/ApeWorX/ape-blast",
     include_package_data=True,
     install_requires=[
-        "eth-ape>=0.7.6,<0.8",
-        "ape-optimism>=0.7.2,<0.8",
+        "eth-ape>=0.8.1,<0.9",
+        "ape-optimism",  # Ape determines the version
         "ethpm-types",  # Use same version as eth-ape
     ],
-    python_requires=">=3.8,<4",
+    python_requires=">=3.9,<4",
     extras_require=extras_require,
     py_modules=["ape_blast"],
     license="Apache-2.0",
@@ -80,7 +80,6 @@ setup(
         "Operating System :: MacOS",
         "Operating System :: POSIX",
         "Programming Language :: Python :: 3",
-        "Programming Language :: Python :: 3.8",
         "Programming Language :: Python :: 3.9",
         "Programming Language :: Python :: 3.10",
         "Programming Language :: Python :: 3.11",