--- a/tests/conftest.py
+++ b/tests/conftest.py
@@ -13,14 +13,14 @@ def accounts():
 
 
 @pytest.fixture
-def base(networks):
-    return networks.base
+def blast(networks):
+    return networks.blast
 
 
 @pytest.fixture(autouse=True)
 def eth_tester_provider():
     if not ape.networks.active_provider or ape.networks.provider.name != "test":
-        with ape.networks.base.local.use_provider("test") as provider:
+        with ape.networks.blast.local.use_provider("test") as provider:
             yield provider
     else:
         yield ape.networks.provider