--- a/tests/test_provider.py
+++ b/tests/test_provider.py
@@ -1,5 +1,5 @@
 def test_basic(accounts, networks):
-    with networks.base.local.use_provider("test"):
+    with networks.blast.local.use_provider("test"):
         a = accounts.test_accounts[0]
         receipt = a.transfer(a, 100)
 
@@ -8,7 +8,7 @@ def test_basic(accounts, networks):
 
 
 def test_get_receipt(accounts, networks):
-    with networks.base.local.use_provider("test"):
+    with networks.blast.local.use_provider("test"):
         transfer = accounts.test_accounts[0].transfer(accounts.test_accounts[1], 1)
         assert transfer.txn_hash
         tx = networks.provider.get_receipt(transfer.txn_hash)