### What I did

before:

```
In [1]: %time import ape_blast
CPU times: user 1.88 s, sys: 228 ms, total: 2.11 s
Wall time: 2.23 s
```

after:

```
In [1]: %time import ape_blast
CPU times: user 8.19 ms, sys: 4.19 ms, total: 12.4 ms
Wall time: 26.9 ms
```

### How I did it

### How to verify it

### Checklist

- [ ] Passes all linting checks (pre-commit and CI jobs)
- [ ] New test cases have been added and are passing
- [ ] Documentation has been updated
- [ ] PR title follows [Conventional Commit](https://www.conventionalcommits.org/en/v1.0.0/) standard (will be automatically included in the changelog)